funds_simulator_api_rest/requirements.txt
fastapi==0.78.0
pydantic==1.9.1
pymongo==4.1.1
python-dotenv==0.20.0
uvicorn==0.17.6
